refactor: rename env to clientConfig
